BackEnd/requirements.txt
fastapi==0.109.2
uvicorn
python-multipart
pillow
jinja2
pydantic==2.6.1
ultralytics
python-dotenv
tensorflow-cpu
numpy<2.0.0
